Create README.md
Skyblock Dupe-Checker Serverwide!

Essential-Dupe-Checker A 1.8.9 Minecraft Forge Mod for Hypixel Skyblock which adds a Dupe Checker to Skyblock

What does it do? This mod adds High-Value Dupe checker into your skyblock experience, It also tracks UUID of items and compare them serverwise to make sure the items you buy are 100% safe and not duped!

Why is this useful? Many players have been wiped before, now this will prevent it. Don't buy anything before Dupe-Checking it!

Is this bannable? Short answer: This mod follows the general interpretation of Hypixel's rules, does not modify packets in any way, and is most likely not bannable, but as with all mods, it is use at your own risk.

Hotkeys: (Configurable in Controls Menu)

/edc on - Activates the dupe checker

/edc off - Deactivates the dupe checker